// fe/src/lib/api/generated/model/activityPageResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { ActivityItemResponse } from './activityItemResponse';

export interface ActivityPageResponse {
  has_next: boolean;
  items: ActivityItemResponse[];
  page: number;
  page_size: number;
}